requirements.txt
# Snowflake
snowflake-connector-python==3.6.0
snowflake-sqlalchemy==1.5.1

# AWS
boto3==1.34.0
botocore==1.34.0

# dbt
dbt-core==1.7.0
dbt-snowflake==1.7.0

# Airflow
apache-airflow==2.8.0
apache-airflow-providers-snowflake==5.1.0

# Utils
pyyaml==6.0.1
python-dotenv==1.0.0
pandas==2.1.4
pyarrow==14.0.1

# Testing
pytest==7.4.3
pytest-mock==3.12.0
